cleaning up in Raybuilder, moving all micromap test-code to designated module
